+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Hosting/INamedPipeProjectInfoDriver.cs
@@ -3,8 +3,9 @@
 
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.Razor.ProjectSystem;
 
-namespace Microsoft.CodeAnalysis.Razor.ProjectSystem;
+namespace Microsoft.AspNetCore.Razor.LanguageServer.Hosting;
 
 internal interface INamedPipeProjectInfoDriver : IRazorProjectInfoDriver
 {
